+++ CHANGES.md
@@ -11,7 +11,7 @@ Apollo 2.4.0
 * [Fix: Resolve issues with duplicate comments and blank lines in configuration management](https://github.com/apolloconfig/apollo/pull/5232)
 * [Fix link namespace published items show missing some items](https://github.com/apolloconfig/apollo/pull/5240)
 * [Feature: Add limit and whitelist for namespace count per appid+cluster](https://github.com/apolloconfig/apollo/pull/5228)
-* [Feature added limit the number of items under a single namespace](https://github.com/apolloconfig/apollo/pull/5228)
+* [Feature add limit for items count per namespace](https://github.com/apolloconfig/apollo/pull/5227)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/15?closed=1)